// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/vehicle-log.model.ts
export class VehicleLog {
    ID: number;
    ActionLogTypeID: number;
    LogTypeAr: string;
    LogTypeEn: string;
    VehicleID: string;
    VehicleCode: string;
    VehiclePlateNo: string;
    VehicleStationId: string;
    VehicleStationName: string;
    VehicleCapacity: number;
    VehicleServiceTypeID: number;
    VehicleServiceTypeEN: string;
    VehicleServiceTypeAR: string;
    DriverID: string;
    DriverName: string;
    DriverMobile: string;
    CreateTime: Date;
    StatusID: number;
    WorkOrderID: number;
    OrderNumber: string;
    CustomerClassesAr: string;
    CustomerClassesEn: string;

}